notification-service-deployment/requirements.txt
aws-cdk-lib>=2.80.0
constructs>=10.0.0
aws-cdk.aws-lambda-python-alpha>=2.80.0a0
boto3>=1.26.0
